requirements.txt
APScheduler==3.10.4
beautifulsoup4==4.12.3
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.3
